MGR: Fix windows compile warnings (Visual Studio 2013).

+++ clientgui/DlgHiddenColumns.cpp
@@ -303,11 +303,11 @@ void CDlgHiddenColumns::OnSize(wxSizeEvent& event) {
 void CDlgHiddenColumns::OnCheckboxClick(wxCommandEvent& event){
     bool bAllOffInGroup, bEnableOK = true;
     
-    int actualCount = m_checkbox_list.size();
-    for (int i=0; i<actualCount; ++i) {
+    size_t actualCount = m_checkbox_list.size();
+    for (size_t i=0; i<actualCount; ++i) {
         std::vector <wxCheckBox*> *checkbox_list = m_checkbox_list[i];
         bAllOffInGroup = true;
-        for (int j=0; j<checkbox_list->size(); ++j) {
+        for (size_t j=0; j<checkbox_list->size(); ++j) {
             wxCheckBox* ckbox = (*checkbox_list)[j];
             if (ckbox->GetValue()) {
                 bAllOffInGroup = false;
@@ -326,14 +326,14 @@ void CDlgHiddenColumns::OnCheckboxClick(wxCommandEvent& event){
 
 
 void CDlgHiddenColumns::OnOK(wxCommandEvent& event) {
-    int actualCount = m_checkbox_list.size();
+    size_t actualCount = m_checkbox_list.size();
     wxASSERT (m_pBOINCBaseView.size() == actualCount);
     
-    for (int i=0; i<actualCount; ++i) {
+    for (size_t i=0; i<actualCount; ++i) {
         CBOINCBaseView* pView = m_pBOINCBaseView[i];
         std::vector <wxCheckBox*> *checkbox_list = m_checkbox_list[i];
         wxArrayString orderArray;
-        for (int j=0; j<checkbox_list->size(); ++j) {
+        for (size_t j=0; j<checkbox_list->size(); ++j) {
             wxCheckBox* ckbox = (*checkbox_list)[j];
             if (ckbox->GetValue()) {
                 wxString name = ckbox->GetLabel();
